requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.7.4
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.6
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
git-filter-repo==2.47.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.1
openai==1.59.7
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pydantic==2.10.5
pydantic_core==2.27.2
pytesseract==0.3.13
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sniffio==1.3.1
starlette==0.41.3
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
